Update patch set 1
Patch Set 1:

I created a local branch with two commits (change default sorting, add togglie for sorting) but was not able to push this branch. Making several commits is in line with the contributor's guide (granularity). But how to do that with Gerrit? It does not make sense to me to push the first commit and then *amend* it with the second. Commits with different concerns should remain separate!?

Patch-set: 1