Update patch set 1
Patch Set 1:

Whenever I got complaints of colleagues about not being able to push multiple commits, then it was caused by one of the commits not having a change id.

Do you have a green overlay at the egit repo icon? If not, please following this from the user guide: "If a repository cloned from a Gerrit server outside of EGit is used in EGit, the clone may not be set up such that EGit recognizes it as a Gerrit repository. Gerrit operations in EGit will not appear in any menus unless you configure your repository first. To do that, open the Git Repositories view and browse down to the Remote that represents the Git repository server you want to use and select Gerrit Configuration...."

Patch-set: 1