Unify handling of "Apply Stash"
EGit had two nearly identical implementations for applying a stash,
once for a node in the git repositories view and once for a commit
editor.

Turns out that the one in the commit editor was general enough to be
used in the repositories view, too. Delete that one and replace by the
one from the commit editor. However, keep the "public" ID: the ID
"org.eclipse.egit.ui.team.stash.apply" had a category and was thus
shown by default in the key bindings preference page, whereas the
"org.eclipse.egit.ui.commit.StashApply" ID had no category and was
thus invisible. Users are thus more likely to already have a key
binding for the old ID. Giving the new ID now a category leads to
the "Apply Stash" command show up twice in the key bindings dialog
because the old ID is not removed from workbench.xmi, even if it is
removed from EGit. (That would be bug 415382...)

This fixes bug 545031 for the "Apply Stash" command, but not yet for
the other commands available in the commit editor.

Bug: 545031
Change-Id: Ie667f964c97d67a5b0e4dfcd4231ee567411995b
Signed-off-by: Thomas Wolf <thomas.wolf@paranor.ch>