Remove unused command ID for "Delete Stash..."
EGit has two command IDs for deleting stashes. The one from the
commit editor's handler had no category, but the other one had no
handler and appeared nowhere, because the handler is bound to the
standard "delete" command.

Use the one with the category for the commit editor's handler. This
enables a user to define a key binding for this handler and invoke
it in the commit editor via the keyboard.

Contrary to the "Apply Stash" case, the handlers from the commit
editor and from the repositories view are not similar enough to
warrant a code refactoring. Maybe if one made the StashedCommitNode
adapt to IRepositoryCommit, but then one has to be careful that no
other commands get enabled inadvertently on such nodes.

Bug: 545031
Change-Id: I5dd3e0e49746babcb825478d16be0eadab5dfa38
Signed-off-by: Thomas Wolf <thomas.wolf@paranor.ch>