Introduce repository groups to Repositories View
In order to allow the user to structure repositories a new node type
'group' is introduced. Groups can be created, renamed and deleted. If a
repository belongs to a group, it is not shown in the tree root.
Assignment is done by having repositories selected while creating the
group or by drag/drop.

Bug: 545123
Change-Id: I7bf62760f0f242f2f72db1e674ed26141964fde4
Signed-off-by: Alexander Nittka <alex@nittka.de>