Introduce repository groups to Repositories View
In order to allow the user to structure repositories a new node type
'group' is introduced. Groups can be created and deleted. If a
repository belongs to a group, it is not shown in the tree root.
Assignment is done by creating the group while having selected
repositories. This is a very rough proof of concept in order to discuss
the basic functionality.

Bug: 545123
Change-Id: Id2965e66c81709653f16f78dc138dca3bf222eea
Signed-off-by: Alexander Nittka <alex@nittka.de>