Update patch set 1
Patch Set 1:

If the problem comes from the overlapping temp and git repositories, a simple fix to try would be to checkout into a sub directory of the workspace, and to create the temp directory as a second sub directory of the workspace.
This should only require that we set a checkout location in the git configuration part, change the shell for the temp, and add the name of the repository directory in the maven, junit and archiving steps.

Do you see any problem with that or shall I give that a try instead? Meanwhile I've also started to rewrite the build from scratch as pipeline. That should allow to reduce the duplication in the many builds. At work I do the builds of all eclipse related products, repositories, branches etc. with a single pipeline definition. :)

Patch-set: 1