Update patch set 1
Patch Set 1:

> If the problem comes from the overlapping temp and git
 > repositories, a simple fix to try would be to checkout into a sub
 > directory of the workspace, and to create the temp directory as a
 > second sub directory of the workspace.
 > This should only require that we set a checkout location in the git
 > configuration part, change the shell for the temp, and add the name
 > of the repository directory in the maven, junit and archiving
 > steps.
 > 
 > Do you see any problem with that or shall I give that a try
 > instead? Meanwhile I've also started to rewrite the build from
 > scratch as pipeline. That should allow to reduce the duplication in
 > the many builds. At work I do the builds of all eclipse related
 > products, repositories, branches etc. with a single pipeline
 > definition. :)

Go ahead and try it. But EmptyOrNewPage should still be fixed to not run up to root. It finds a .project that doesn't belong to the test Eclipse's workspace. So it should stop at the repo root or at the workspace root, whatever is hit earlier. (Uhh... what if the user tries to share a project into a .git repo that is _inside_ the workspace? Allowed or not?)

Patch-set: 1