Reset only conflicting paths in a checkout conflict
The two checkout conflict dialogs show only the conflicting files,
but their "Reset" button then performed a reset --hard, also discarding
changes on other files not shown in the dialog. This was highly
unexpected.

Change that reset button to only discard changes on the files shown
in the dialog, i.e., the ones that caused a conflict in the first
place. Use "Discard Changes" instead of "Reset" for the button text.

Add a new test that verifies that non-conflicting changes are preserved.

Bug: 545471
Change-Id: I4bc7045cc98b04fe6254cffe1fc44a64bb6e2d88
Signed-off-by: Thomas Wolf <thomas.wolf@paranor.ch>