Integrate smart import wizard fragment into EGit UI bundle
Move the class into the main bundle now that the EGit baseline
is a version that supports smart import. Add the wizard extension
in plugin.xml. Move the test class into the EGit UI tests and
rewrite it to use a local test repo instead of trying to clone
JGit over HTTPS.

Remove the smartimport fragment and the ui.importer.tests project.
Add the update clause to the p2.inf of EGit UI to ensure that p2
removes the fragment when EGit UI is updated. Also remove the
smartimport fragment from the repository.

Change-Id: I6b9c3138da25d77995ef0acd9168ddccc3663568
Signed-off-by: Thomas Wolf <thomas.wolf@paranor.ch>