Add code search to Git Code Search (Prototype)
This is a proof of concept for adding code search into Git Search. This
type of search is not covered by File Search as repositories need not be
present the workspace. The commit is intended as the starter for a
discussion whether this feature should be considered at all.

Only the head version of the working tree is checked for the pattern
(checked out branch or all branches depending on search settings).
In this minimal implementation a search hit causes a sysout of
repository name, branch name and file path.

Change-Id: Ida3e7d5dfa62d3d56648a35f782860ddffa72e00
Signed-off-by: Alexander Nittka <alex@nittka.de>