Update patch set 1
Patch Set 1:

Hi Thomas, 
Thanks for the feedback. I think it does not make sense to use a switch if already right now the choose to use a simple string comparison is wrong. The change to a switch is a semanticaly change to a decision „one of <n>“. As soon as you need another processing step like tolowercase() or something like this then there is no way in java to reflect this other than a if-then-else chain (or insert more code to compute a variable for this decision).
I was not aware of that it is so wrong to use a switch in this case but I think I would rather rely on your experience regarding this code. Should I close this request? Can you do it?

Patch-set: 1
Label: Verified=0