Update patch set 1
Patch Set 1:

Just to be sure that I understood the formal requirements correctly:

The two CQs are for JGit, not for EGit. This just adapts the EGit target platforms such that one can have both EGit and JGit in the same workspace. EGit does not include sshd in its repository, it gets included in the EGit update site because we re-distribute the JGit features. Using and redistributing another Eclipse project doesn't need a CQ. So this change would not be required to mention the CQs, and we don't need separate CQs for EGit either.

Is that correct? If not, then why don't we need CQs for EGit?

Patch-set: 1