History: new operation to check out files from a commit
Add a new operation in the CommitFileDiffViewer to check out file
revisions shown into the working tree. Previously the user would
have to navigate to the file in the package explorer (or project
explorer) and use "Replace with...->Commit" there, which opens a
dialog to choose the commit. Or use "Open This Version" and then
copy-paste the text. Both are rather round-about ways to achieve
this. Newly there's a "Check Out This Version" command in the context
menu that enables the user to do this directly.

The same operation is also available in the outline view of the
unified diff page of the commit viewer.

The code checking for potential loss of uncommitted changes and
for running launches and then asking the user to confirm the
operation has been factored out of DiscardChangesActionHandler
into CommandConfirmation.

Bug: 362907
Change-Id: Ie237e24b9e78b888b0de816d39c1c43ca44c400a
Signed-off-by: Thomas Wolf <thomas.wolf@paranor.ch>