Update patch set 4
Patch Set 4:

(3 comments)

Good start, but needs more work. Editable logic for the tag message is inverted.

Should be able to move lightweight tags. Allows to turn an annotated into a lightweight tag, but not vice versa. Setting the tag message read-only when "lightweight" is chosen is a bit confusing; there's no visual feedback. Perhaps additionally clear the text area (saving the text, so that it can be restored if "lightweight" is switched off and the user didn't select a different existing tag in the meantime).

Patch-set: 4