Fix NPE in RepositoriesViewPropertyTester
There can be a race condition when first getting the list of all child
paths and then iterating the paths to get the ref again, since the ref
can have been deleted in the meantime by another thread.

Avoid the race condition by using just a single interaction with the ref
database.

Bug:548434
Change-Id: Ie665d4b1c7c49668550a5f75341695b1a24c52a2
Signed-off-by: Michael Keppler <Michael.Keppler@gmx.de>