Update patch set 9
Patch Set 9:

> I tried to change the test in the latest patch as you said but the
 > assertion fails anyway

No, TagActionTest passed. But TagOperationTest failed since it expected an exception for the NO_CHANGE case.

I've fixed that, and have made the TagActionTest a bit more strict such that it not only checks for the ref being != null but actually tests that the tag points to the right commit, is annotated or lightweight, and if annotated has the right message.

Patch-set: 9
Label: Verified=0