Update patch set 10
Patch Set 10:

your test fails even locally for me as my test did. I don't understand why if I do the operation manually and I check the git repo the annotated tag is correctly converted but the test fails. I don't understand if the test is correct but implementation is wrong or it's just a test problem.

Patch-set: 10