Update patch set 11
Patch Set 11:

I think I found the problem. It's reproducible also in manual testing. GTK only, apparently.

When SWTBot selects a TableItem, it focuses the table first. Focusing the tag table triggers an OpenStrategy that somehow selects the first table item. There's a selection listener on the table that sets the tag name text field, which acts as a filter on the table. So the moment the tag name field is set, all items from the table but the first one disappear. Then SWTBot tries to select the "SomeTag" item (the second one, gone now), which doesn't work. So we end up with the "SomeLightTag" being selected and then being changed by the test.

To reproduce manually: open the "Create Tag..." dialog on the JGit repo (or any other repo having more than one tag). Initially shows all tags, tag name text box has focus. Now TAB-navigate through the fields, using Ctrl-TAB to get out of the tag message field. Keep navigating until the table gets the focus. Oops, all tags but the first one are gone.

On Mac, manual reproduction is not possible (TAB navigation never focuses the table), and the problem doesn't appear during the test either. Why a focus event on the table selects the first element on GTK is unclear to me.

All the other tests work because they either want to select the first tag in the table anyway, or they use the tag name field to choose the tag. testConvertAnnotatedTagIntoLightWeight() is the only test that wants to select the second item in the table, and does so via the table.

Patch-set: 11
Label: Verified=0