Fix showing selected refs or commits in history view
When the history view shows only the current branch ("Show all
branches and tags" _off_), selecting some other branch in the git
repositories view might not show the selected branch. It would be
shown only if that branch happened to be merged into the currently
checked out branch. The same could be observed when using "Show In
History" on a commit in the results of a "Git Search": the commit
would only be selected if it was contained in the branch currently
checked out.

Change this by always considering such a selected branch or commit
as an additional start point for the RevWalk. This enables the user
to make the history view show exactly two branches (plus all those
merged into these two), and also correctly selectes commits not on
the current branch.

Bug: 360868
Change-Id: Ib2e071f40c57b06a0ce0a51defac3fe30f481e03
Signed-off-by: Thomas Wolf <thomas.wolf@paranor.ch>