Update patch set 2
Patch Set 2: Code-Review+1

(2 comments)

Tested. Seems to be working nicely.

I noticed that we have yet another bug with the show in history: If the currently selected branch is already the input, then nothing is done (which is generally right). but if the branch has been scrolled out of view, we simply return without scrolling to the top of that branch. Nevertheless, that seems completely independent of both your history changes, so I may create a separate new bug, if you want.

Patch-set: 2
Label: Code-Review=+1