Update patch set 1
Patch Set 1:

> Are you aware of the identity comparison in GitHistoryPage line
 > 1825? Would this benefit from a similar change? That line is
 > explicitly meant to filter away repeated similar inputs.

Is it? I thought it was meant to handle the case where getMostFittingInput() didn't change the input object. I think == is correct there.

What we could do in a follow-up change is to change HistoryPageInput.equals() to handle IResource and FileNode/FolderNode from the repo view as equal if they refer to the same repo-relative path in the same repo. The == on the repos is also a bit fragile; it works because we always use the same repo instance from the RepositoryCache, but maybe comparing repo.getDirectory() (with appropriate null checks if one of the repo fields is null) might be better.

Patch-set: 1