Update patch set 3
Patch Set 3:

> With this change fetched, and "sync with selection" active in
 > history view, I get exceptions when selecting very old tracking
 > branches in platform.ui, like "origin/perf_33x".

Cannot reproduce on Mac. The only thing I see is that if the history is incomplete (cut short before the commit the old branch points to), the branch is _not_ selected. That, however, is IMO acceptable. Better to respect that limit than trying to keep going until the wanted commit is found. The user probably had his reasons to set the limit...

Patch-set: 3
Label: Verified=0