Put a "presentation" button in the unstaged viewer's toolbar
Adds a button to switch the file presentation in both the unstaged
and staged viewers between list mode, tree mode, and compact tree
mode to the toolbar of the unstaged viewer, where we already have
the button to switch the sort order that also affects both viewers.
Re-uses the actions defined for the view menu. The button changes
its icon depending on the mode chosen.

A first implementation made selecting the button cycle through
the presentations, but that introduced a usability problem: when
clicked and the presentation switched to list mode, the two buttons
for "expand all" and "collapse all" would be hidden. The mouse would
then suddenly be over the "stage selected files" button. A user who'd
mistakenly click twice might inadvertently stage things he hadn't
wanted to stage.

Thus this implementation now simple shows the drop-down menu also
when the button is selected, not only when the arrow is selected.
When the user selects an entry from the drop-down menu and the
toolbar changes, the mouse won't be over any of the buttons.

Note that the staging view reacts only upon the next refresh or
repository switch on changes of "Maximum number of lines" preference
that governs the automatic switching to tree mode. Thus when this
preference is changed to a higher value and the view currently is in
that forced list mode, the "list" menu item may remain disabled until
the next refresh. That issue is unrelated to this change, though. (But
it is made more visible.)

Bug: 549269
Change-Id: I347941440ec411eaa21b4e014ebc44893e6c28ce
Signed-off-by: Thomas Wolf <thomas.wolf@paranor.ch>