Add show first parent only button to history view
Add an action to the GitHistoryPage to show first parent only with some
extra code to allow the setting to be sticky to one repository.
Also add getter and setter for GitHistoryPage.currentRepo, so a repo
change can be detected.
Add the action to the toolbar and menu.
Add the corresponding preference to the preference page.
Also add a new icon for the action.
Finally implement, that the corresponding preference sets first parent
on the RevWalk.

Bug: 549424
Change-Id: I548ef25f8c13661fadf725952bc25d2f481c7a43
Signed-off-by: Tim Neumann <Tim.Neumann@advantest.com>
Also-by: Tetiana Pliekhova <tetiana.pliekhova@advantest.com>
Signed-off-by: Andrey Loskutov <loskutov@gmx.de>