Update patch set 1
Patch Set 1:

First: GREAT. I finally see my branch history!!!

I've tried this on our monster repo and the new mode must be default *for our* IDE.

Because of this: please add the same option in the Team -> Git -> History view preference page. This way I can set it in Oomph / make a product default for us.

Last point (before looking on the code): the "x" on the icon must be not that saturated, it is too dark. Also I think it can be moved few pixels to the right so that the left part of the merge arrow is fully shown.

Patch-set: 1
Label: Code-Review=0