Fix a few instances where repositories are not closed
In the staging view, ensure the fake RepositoryNode is cleared when
no longer needed.

In ExistingOrNewPage, ensure the newly created repo is closed.

In LinkHelper, look up the repository from the repository cache
instead of creating a new Repository instance that then cannot be
closed because it my be referenced by the RepositoryTreeNode returned
and that may be needed later.

Change-Id: I7d255df59038167b17796735cc36b542d0771d36
Signed-off-by: Thomas Wolf <thomas.wolf@paranor.ch>