History view: shorten view toolbar
Combine the four filter buttons into one single button with a
drop-down menu. Selecting the button itself cycles though the
filters. Use the new FilterAction also in the view menu; to be
able to do so implement ToolBarMenuAction.getMenu(Menu) and rename
the class to DropDownMenuAction.

Also, as long as one only updates the image of an action it isn't
necessary to update the whole toolbar. The framework listens for
such changes and updates the corresponding ToolItem or Menu entry
already.

Change-Id: If7fb1e3199cd1721a4455bac10925fc2922c5ba8
Signed-off-by: Thomas Wolf <thomas.wolf@paranor.ch>