Update patch set 2
Patch Set 2:

Showing the current state in the tooltip is a bit strange, too, since most other tooltips tell what's going to happen. Maybe a fixed tooltip for the button "Change the file scope for the history" might be better.

Or maybe a combined tooltip: "Change the file scope for the history. Current: " + tooltip of current action when enabled, just "Change the file scope for the history" when disabled.

Patch-set: 2