Update patch set 1
Patch Set 1:

> > The setting should be persisted per repository; that would make
 > it
 > > even more useful.
 > 
 > Can you point me into a direction on how to do that?

I would have to try to be sure, but here's an idea: store a list of git directories in the preference. The preference is changed only by that button, so you can read the list once and keep it in memory. Then set first-parent only for repositories whose git dir is in that list.

When a repository is removed, you'd probably also want to remove the git dir from that preference. There's a change by Matthias already to clear the history when a repo is deleted, see https://git.eclipse.org/r/#/c/146471/ . That preference listener could be used to remove the git dir from the "first-parent" preference.

Probably would want a global default setting, too. User preference, settable on the preference page. "Use --first-parent by default". In your product, you could set that global pref to true so that it is enabled out of the box without having to know where the user's local clone is.

Other users can leave that global pref false; --first-parent will then be used only if they click that button while the history of a repository is shown.

Patch-set: 1