Handle Java templates initialization error silently
If the additional variable for the Java templates cannot be registered,
only log this error to the platform log, but do not show an error
dialog.

Depending on the (currently unknown) root cause of the error, this might
happen on every start of eclipse, and users will be annoyed by an error
message that is shown on every eclipse start.

Bug: 549206
Change-Id: Ife0bf013c9f5666218447426f03002b87687f83c
Signed-off-by: Michael Keppler <Michael.Keppler@gmx.de>