Update patch set 7
Patch Set 7:

> Since gerrit is broken once again right now, I cannot test this
 > locally. However, I have a question: Is it generally useful to have
 > all 4 combinations of "first parent only" x "show all branches and
 > tags"? I.e. reading the documentation I believe that "first parent
 > only" AND "show all branches" both enabled does not make sense,
 > right?

Hmm, I've tried now all 4 combinations of those two buttons, and depending on the code / history they all can show different results and so potentially be useful for some cases.

 > If that assumption is true, then I suggest to have a dropdown
 > action with 3 items instead of the 2 toolbar actions, and to cycle
 > through them like in Thomas content filter change at
 > https://git.eclipse.org/r/#/c/146411/


3 items instead of 2 toolbar actions? Do you mean 2 or 4 (either/or state or all combinations)? Or what are "3" items?

Regarding the proposal to combine them - not sure how one could meaningfully name the states and which icons could be appropriate. That could be confusing. 

May be we just don't need to put the new button on the toolbar but keep it in the view menu? 

I think this could be less confusing to see similar looking buttons next ro each other (menu shows action names, so it is easier to understand the difference).

Patch-set: 7
Label: Verified=0