Update patch set 8
Patch Set 8:

(16 comments)

I think from the functional / UI point of view we are at right state now, it works as I would expect.

I have some smaller issues with the code below, but nothing critical.

General comment: Thomas, is there any reason why we introduce here repository specific settings for *one* preference, and have nothing similar for others? 

This seem inconsistent, if some history specific settings are per repo (I believe only this one), and some are not, but they *could* be, if we apply same logic. Looking on the "History" prefs page, "Branch sequence" and "Tag sequence" are the settings that we *must* disable for our monster repo to avoid minute long pauses - but they are absolutely fine and useful for smaller repositories. Similar for the "max numbers of..." etc preferences. Should we apply "per repo" logic for other settings and create new enhancement request for that?

Patch-set: 8
Label: Verified=0