Update patch set 11
Patch Set 11:

Please re-add the button in the view toolbar. I think having some visual indication whether first-parent is on or off is a necessity. If we don't have that button, we'd need some other way to indicate the state, without wasting space. Maybe in the view description, but that feels contrived. Better just have the button.

Patch-set: 11