Update patch set 11
Patch Set 11: Code-Review-1

The preference handling sometimes gives still surprising results, but I guess there's not much to be done about it. Ideally we'd have tri-state buttons...

Start with the default preference set to true. Visit some repo. Uncheck the repo-specific first-parent option. Toggle the global default: no effect, since we set a repo-specific setting. Toggle the global default again; again no effect, as expected. Check the repo-specific setting again. The last removes the repo-specific property, so if you now set the default preference to false, that will also have an effect on the repo shown. Even though the user explicitly switched first-parent on for that repo.

Maybe the whole thing would be easier to understand if an explicit click for the repo-specific setting always set the repo-specific setting to true or false, but never removed it. Don't know... @Michael, your opinion?

(BTW, a good repo to see the effects of first-parent is github.com/git/git :-)

Patch-set: 11
Label: Code-Review=-1