Update patch set 2
Patch Set 2:

I think this should have a test. I know the code only deleted the children of the folder node, not those of the real on-disk directory, but to catch regression we should have a test for this. A folder x with two children a and b, only a staged. Delete invoked on folder x. Expected result: on disk x/a still exists, and x/b is gone.

What should happen if x is empty after the deletion? Should x itself also be deleted?

What happens if x in the staging view is not expanded? Does the user have a notification which files will be removed? I guess this makes it rather easy to accidentally delete files one didn't want to delete. (Just think of someone mistakenly having selected a folder.)

Patch-set: 2