Update patch set 2
Patch Set 2: -Code-Review

Thomas, the handling of files or folders on disk did not change at all. E.g. all the elements that we deal with are StagingEntries for _files_, before and after my change. Selecting a folder staging entry only adds its (file) child entries from the view, the same as if you had selected these child nodes manually. The selection list dealt with in the code is exactly the same, whether you select the folder node, or all of its child nodes. My change only allows to create that selection list differently.

Due to only dealing with files, we cannot and do not delete empty directories, neither before nor after my change.

The staged and unstaged entries are in 2 separate trees, therefore a staged entry cannot accidentally be added to the deletion list in the unstaged view (what you suggest with the x/a and x/b files). Also that is independent of my change.

To summarize, the possible actions on the staged/unstaged file nodes are exactly the same as before. Do you agree it is not necessary to have additional testing for the effects of the file actions?

Patch-set: 2
Reviewer: Gerrit User 440 <440@97ee7c02-f12f-4043-b43e-dea463d88b31>
Label: Code-Review=0