Update patch set 3
Patch Set 3:

> > Wouldn't that break the separation between core and UI?
 > 
 > This is gitflow.ui, so I think it'd be OK it these handlers used
 > BranchOperationUI. (I should have written "handlers" instead of
 > "operations".)

Right, it would be OK to use BranchOperationUI in the handlers, though the BranchOperation is triggered in the handlers. I would need to turn that off via some flag when running the operations from the handlers, and then duplicate the logic using BranchOperationUI. This would feel like the beginning of a "spaghetti" to me. Or did you have something different in mind?

Patch-set: 3