Update patch set 3
Patch Set 3: Code-Review-1

I'm not sure this is the right approach. I also don't quite see what it brings. What exactly was the use case that prompted this?

Open history view with "Link to selection" on.

Open an editor on some tracked file. Go to the repo view and select another folder from the same repo. Go back to the editor.

History now shows the history of the file in the editor. Now I want to navigate elsewhere in the repo view. Activate repo view.

History view now shows the history of that folder. So I've lost the history I wanted to see.

If you want to keep a history, switch off "link to selection" in the history view, or pin it.

Patch-set: 3
Label: Code-Review=-1