[releng] Update target platform & SystemReader to JGit changes
During the fixes in JGit 5.1.9 the JGit target platform was changed,
and SystemReader got two new abstract methods.

EGit doesn't build anymore after these changes. Update the target
platform to match the one in JGit to enable building both EGit and
JGit in the same workspace, and add implementations for the two new
methods in the EclipseSystemReader.

Change-Id: Ic1392b8e140d2be118226567a71be62ba833477d
Signed-off-by: Thomas Wolf <thomas.wolf@paranor.ch>