Update patch set 2
Patch Set 2:

> I definitely see an improvement compared to before. But I'm
 > wondering about using a different strategy: Instead of restoring
 > the current top commit at its top position, can't we restore the
 > currently selected commit at its index? I.e. such that the selected
 > commit remains the same and does not move.
 > 
 > My thinking is: I select some commit in the middle of the history
 > view (with all branches disabled). Then I enable all branches. I
 > expect the selected commit to remain where it is, not the top
 > commit. Actually when testing your change I first thought of a bug,
 > because I only looked for the selected commit...

This is a cool idea, but may be extremely difficult to implement. There may be multi-selections, which may or may not contain the commitToShow, which may or may not be visible already. Some commits from that multi-selection may disappear. So besides having to figure how many rows are visible and which of the selected items are visible, one then also has to compare that to the selection _after_ the update and figure out which previously selected visible items remain selected, and finally whether to keep the index of the first of these items, or the top index all the same (incremental loading).

We could try to improve this along these lines, but right now I have no time for this.

Patch-set: 2
Label: Verified=0