Update patch set 3
Patch Set 3:

> What are your settings? Show all branches? Additional refs?

And "Follow selection"?

In an Eclipse with a somewhat older EGit without all these changes in the history view, with follow selection off, show all branches and additional refs on: when I scroll down in the history and select a commit, then do a fetch from upstream in the repo view, the git history also does not scroll to the top. It does update; the fetched commits are there when I scroll manually to the top.

Patch-set: 3
Reviewer: Gerrit User 44628 <44628@97ee7c02-f12f-4043-b43e-dea463d88b31>
Label: Verified=0