Update patch set 3
Patch Set 3:

You are probably right, that this is related to settings. I have to check the exact settings at work, but I remember that I started using several repo view filters and to not show additional refs in history, so that might be it.

With that explanation, I think there is no need to further investigate. Thanks for looking into it again.

Patch-set: 3