Cache more repository state for re-use in the UI thread
Many EGit commands depend on some repository state determining their
enablement. Eclipse re-evaluates activation/enablement expressions or
handlers' isEnabled() method very frequently. For globally available
handlers, on every selection change; for others for instance when
they are included in some (context) menu.

JGit operations to determine repository state access the file system
(typically config files, or the packed_refs file, or loose refs), and
check every time whether the file has been changed since the last
access. In the case of a repository config file, this check is even
done for the user and system config files, too. If JGit determines
that a file has been changed it reloads it, perhaps even multiple
times depending of file timestamp resolution.

Thus accessing the repository config file or determining the currently
checked out branch are expensive file system operations. Handler
enablements are evaluated on the UI thread, so EGit did *a lot* of
expensive file system accesses on the UI thread, which slows down
Eclipse startup and context menus in particular.

Introduce a cache for the most important values of a repository that
are used in property testers and in handler activation and enablement
code. The cache is cleared whenever the workbench selection or the
"menu selection" changes to avoid serving stale values. When Eclipse
then re-evaluates enablements, only the first expressions will actually
access the file system, while all later re-evaluations due to a
selection change will use the cached values.

Caching values until the next selection change is a valid strategy.
When a series of enablements is re-computed, these re-computations
are all based on the same selection, thus using the repository state
as it was at the beginning is fine.

This greatly reduces the number of times EGit accesses the file system
on the UI thread and generally makes the EGit UI *much* snappier.

This cache should be used only for evaluations that are known to
happen synchronously in the UI thread in response to a selection
change. This includes property testers and isEnabled() on handlers.

Bug: 544600
Change-Id: Idd369b50d8b4bd6dd68b30d640bbc13f5a5976c5
Signed-off-by: Thomas Wolf <thomas.wolf@paranor.ch>