Update patch set 2
Patch Set 2:

This is definitely something for early 5.6.0, not 5.5.0. I see thousands of calls to Repository.getConfig(), each resulting in three calls to FileSnapshot.isModified() when running EGit. More calls to FileSnapshot.isModified() come from getting HEAD. And all these calls are on the UI thread! (There's thousands more in the decorator background threads, but those are less critical.)

All these calls on the UI thread come from handler re-evaluations when the selection changes.

This change reduces these thousands of call dramatically.

Using a child Eclipse with the Java perspective, no open editors, no git views open, and the EGit projects in the workspace and shared with the EGit clone, no tree open in the package explorer (only the projects visible) performing the following steps:

1. Start Eclipse
2. Select org.eclipse.egit.ui
3. Open context menu
4. Open the "Team" sub-menu
5. Dismiss the context menu, select org.eclipse.egit.core
6. Navigate to src/org.eclipse.egit.core/EclipseAuthenticator.java
7. Select the file
8. Double click the file; let the editor open
9. Shutdown

I get the following counts for calls on the UI thread without this change:

4.6  target: 294 calls to getConfig(), 1322 calls to isModified()
4.10 target: 271 calls to getConfig(), 1163 calls to isModified()

Same steps from the same initial situation, but this time with the git repositories view initially open and containing 15 repositories (one of which is the egit repo):

4.6  target: 1846 calls to getConfig(), 6104 calls to isModified()
4.10 target: 1498 calls to getConfig(), 4958 calls to isModified()

With this change, the numbers are more reasonable:

First scenario (no git views open)

4.6  target: 23 calls to getConfig(), 107 calls to isModified()
4.10 target: 11 calls to getConfig(), 53 calls to isModified()

Second scenario (git repo view open)

4.6  target: 55 calls to getConfig(), 236 calls to isModified()
4.10 target: 37 calls to getConfig(), 155 calls to isModified()

Makes the UI react noticeably faster even on my Mac. I guess on Windows the effect might be even more pronounced.

For the decorators I don't see a solution yet, but maybe the same cache could be used if it's also cleared on index diff changes. Our decorators are slow for the same reason, but at least they're not on the UI thread.

Patch-set: 2