Cache the current head for CommitGraphTable.doPaint()
doPaint() called input.getHead() every time. This is a fairly expensive
file access occurring on the UI thread for every row. At least, it will
check whether HEAD and the ref it may point to have been modified.

Reduce the number of such calls by caching the head Ref until a new
input is set. This is valid since the table will be re-loaded anyway
when a repository change is detected.

Bug: 544600
Change-Id: Iea994a0f426a574c26501f4151bc568461dc8aaf
Signed-off-by: Thomas Wolf <thomas.wolf@paranor.ch>