Eliminate exactRef(HEAD) on the UI thread in the history table
CommitGraphTable.doPaint() called input.getHead() every time. This
is a fairly expensive file access occurring on the UI thread for
every row that is painted. At least, it will check whether HEAD
and the ref it may point to have been modified.

Eliminate these calls from the UI thread completely by doing it
once in the SWTWalk, which runs in the background. Once determined,
the head Ref can be obtained through the SWTCommit.

The CommitGraphTable (or rather, the SWTPlotRenderer) uses the head
Ref *only* to bolden the branch that is HEAD.

Bug: 544600
Change-Id: Iea994a0f426a574c26501f4151bc568461dc8aaf
Signed-off-by: Thomas Wolf <thomas.wolf@paranor.ch>