Update patch set 2
Patch Set 2:

I suggest to also enable all checkboxes in the PDE>DS Annotations page in the project properties of this plugin. That way we get a modified service declaration file, if code and service declaration go out of sync (even though I don't imagine this happening). Enabling those checkboxes should lead to pde.ds prefs in the .settings directory and a pde.ds builder in the .project file.

We use that policy at work, and it rescued us several times when people didn't correctly modify both sides of the service.

The change as such is completely fine.

Patch-set: 2
Label: Code-Review=0