Update patch set 5
Patch Set 5:

> How about cherry-picking these 2 changes to stable-5.5 to include
 > them in 5.5.0 ? They are improving performance considerably. I am
 > using these patches since you uploaded them for review and didn't
 > come across any issues.
 > 
 > What do you think ?

It should be safe, but personally I prefer in general to do such larger refactorings early in a new release, not late in an upcoming one. Neither you nor me use Windows. It doesn't fix a regression either. 

But if you want to have it 5.5.0, it's fine by me.

Patch-set: 5