Update patch set 6
Patch Set 6:

nice :-)

I think it's a good idea to have predefined common entries which can be selected in the drop down of the toolbar button. In the "Ref filter..." dialog I would suggest to show these predefined (and not removable) patterns with grey text color to make clear they are predefined and cannot be removed. And I would remove the text "(preconfigured)" since the different text color already shows these entries are different.

Maybe we could add content assist when the user edits an entry in the ref filter dialog which proposes predefined entries and the list of existing refs matching the text which was already entered. 

In addition we could add actions on branches/tags shown in the repositories view to select/deselect refs to be used in the history view's ref filter. And we could add menu actions on the ref widgets displayed in the history view to enable a more intuitive point and select way to adjust the ref filter.

The toolbar button can toggle between 2 filter settings. But it does not change appearance between these 2 states. I think it should appear like an unpressed toggle button when only showing HEAD and it should appear like a pressed toggle button for the other state.

The "Add Ref..." button in the Ref Filter dialog seems not to have any effect yet.

Patch-set: 6