Update patch set 1
Patch Set 1:

The ResourceRefreshJob can also be switched off by various preferences, and it's tied (mistakenly, I believe) to the repo change scanner being active. If we start relying on this job, then there shouldn't be a way to disable it.

Also, it's a little bit disconcerting that this job is in UI, but all the calls to refreshValidProjects() are in core.

ProjectUtilTest.testRefreshValidProjects() makes no sense anymore.

Patch-set: 1
Label: Code-Review=0