Update patch set 1
Patch Set 1: Code-Review-2

As expected this has nasty side-effects. It appears to work well for xtext, but it breaks one test and I can even reproduce that with a conflict in workspace on a plain text file. Moreover, it produces an exception on a text file not in the workspace.

So this needs better approach.

Patch-set: 1
Label: Code-Review=-2