Update patch set 6
Patch Set 6:

> I think it's a good idea to have predefined common entries which
 > can be selected in the drop down of the toolbar button. In the "Ref
 > filter..." dialog I would suggest to show these predefined (and not
 > removable) patterns with grey text color to make clear they are
 > predefined and cannot be removed. And I would remove the text
 > "(preconfigured)" since the different text color already shows
 > these entries are different.

I will do the coloring. But I think more discussion is needed before removing the "(preconfigured)" (see comment of Thomas)

 > Maybe we could add content assist when the user edits an entry in
 > the ref filter dialog which proposes predefined entries and the
 > list of existing refs matching the text which was already entered.
 > 
 > In addition we could add actions on branches/tags shown in the
 > repositories view to select/deselect refs to be used in the history
 > view's ref filter. And we could add menu actions on the ref widgets
 > displayed in the history view to enable a more intuitive point and
 > select way to adjust the ref filter.

I will look into this. Do you think any of these are required for this to be a "good" contribution which can be merged? Otherwise I'd propose to do these in a follow up change.

 > The toolbar button can toggle between 2 filter settings. But it
 > does not change appearance between these 2 states. I think it
 > should appear like an unpressed toggle button when only showing
 > HEAD and it should appear like a pressed toggle button for the
 > other state.

I totally agree and already tried to do this. But as far as I understood, this is currently not possible with org.eclipse.jface.action.Action, because it can either have the style AS_DROP_DOWN_MENU or AS_CHECK_BOX, not both.

 > The "Add Ref..." button in the Ref Filter dialog seems not to have
 > any effect yet.

This is weird. It should open a Ref Selection Dialog for the current repository. We need to look into this. But I don't really know how to start.

Patch-set: 6