Update patch set 6
Patch Set 6:

> It seems that "Set defaults" in the ref filter dialog depends on
 > the current toggle state and sometimes selects only HEAD. I think
 > it should always set the predefined settings "refs/tags/**:HEAD:refs/heads/**:refs/remotes/**"

Actually it should always restore to the state where only HEAD is selected. Did you experience a situation where this is not the case?
Reasoning: This is also the default with the old system (toggle between HEAD and all branches and tags).
The string "refs/tags/**:HEAD:refs/heads/**:refs/remotes/**" is only the default for the *last* selected refs, so the state to which is returned, if the toolbar button is pressed, while only HEAD is selected. This is also currently never set to default, will need to look into when to set this to default.

Patch-set: 6