Update patch set 7
Patch Set 7:

But we have a problem here. For instance ResetOperation.execute() is run inside a WorkspaceOperation, so it holds the workspace lock. The RRJ cannot refresh because the lock is held and blocks. Deadlock. You can't even shut down Eclipse because Activator.stop() will join on RRJ, but RRJ hangs waiting on the workspace lock, which is still held by the ResetOperation waiting for RRJ.

Perhaps using the job to do the refreshes isn't the right way. Maybe better to factor out the actual refresh logic and attach a (synchronous) listener for WorkingTreeModifiedEvent to the repo in each operation.

I also tried a cherry-pick. My test repo has one file only, and the cherry-pick produced a conflict. I didn't get a WorkingTreeModifiedEvent at all. That needs to be examined. I would have expected one.

Patch-set: 7